move common AWS related classes into a separate module
