+++ aws-common/src/main/java/io/druid/common/aws/AWSCredentialsConfig.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package io.druid.storage.s3;
+package io.druid.common.aws;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
