+++ extensions/s3-extensions/pom.xml
@@ -37,6 +37,12 @@
             <groupId>io.druid</groupId>
             <artifactId>druid-api</artifactId>
         </dependency>
+        <dependency>
+          <groupId>io.druid</groupId>
+          <artifactId>druid-aws-common</artifactId>
+          <version>${project.parent.version}</version>
+        </dependency>
+
         <!-- override jets3t from hadoop-core -->
         <dependency>
             <groupId>net.java.dev.jets3t</groupId>
@@ -46,10 +52,6 @@
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
             <artifactId>log4j-1.2-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk</artifactId>
         </dependency>
         <!-- override httpclient / httpcore version from jets3t -->
         <dependency>
