+++ pom.xml
@@ -56,6 +56,7 @@
         <module>server</module>
         <module>services</module>
         <module>integration-tests</module>
+        <module>aws-common</module>
         <!-- Non-default modules -->
         <module>extensions/cassandra-storage</module>
         <module>extensions/hdfs-storage</module>
