+++ server/pom.xml
@@ -38,6 +38,11 @@
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
+            <groupId>io.druid</groupId>
+            <artifactId>druid-aws-common</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
             <groupId>com.metamx</groupId>
             <artifactId>http-client</artifactId>
         </dependency>
@@ -60,10 +65,6 @@
         <dependency>
             <groupId>org.glassfish</groupId>
             <artifactId>javax.el</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.curator</groupId>
